Apply changes for Nursery, Vehicle and Weighbridge module

+++ tests/UiTests/NurseryTests/15_NurseryTransferRequisition.spec.js
@@ -1,8 +1,14 @@
 import { test } from "@utils/commonFunctions/GlobalSetup";
+import { allPhases } from "@utils/data/uidata/globalData.json";
 import LoginPage from "@UiFolder/pages/General/LoginPage";
 import SideMenuPage from "@UiFolder/pages/General/SideMenuPage";
 import editJson from "@utils/commonFunctions/EditJson";
-import { checkLength } from "@UiFolder/functions/comFuncs";
+import {
+  setCurrForm,
+  setCurrPhase,
+  checkLength,
+  throwTestFailMsg,
+} from "@UiFolder/functions/comFuncs";
 import {
   validateFormValues,
   validateDBValues,
@@ -17,7 +23,8 @@ import {
 
 import {
   NurseryTransferRequisitionCreate,
-  NurseryTransferRequisitionEdit,
+  NurseryTransferRequisitionEdit1,
+  NurseryTransferRequisitionEdit2,
   NurseryTransferRequisitionDelete,
 } from "@UiFolder/pages/Nursery/15_NurseryTransferRequisition";
 
@@ -28,6 +35,7 @@ let sideMenu;
 let createValues;
 let editValues;
 let deleteSQL;
+let phaseCount = 0;
 const sheetName = "NUR_DATA";
 const module = "Nursery";
 const submodule = null;
@@ -36,9 +44,13 @@ const keyName = formName.split(" ").join("");
 const paths = InputPath[keyName + "Path"].split(",");
 const columns = InputPath[keyName + "Column"].split(",");
 
-test.describe.serial("Nursery Transfer Requisition Tests", () => {
+test.describe.serial(`${formName} Tests`, () => {
   // ---------------- Before All ----------------
   test.beforeAll("Setup Excel, DB, and initial data", async ({ db, excel }) => {
+    // Change Current Form and Phase
+    await setCurrForm(formName);
+    await setCurrPhase(allPhases[phaseCount]);
+
     // Load Excel values
     [createValues, editValues, deleteSQL, ou] = await excel.loadExcelValues(
       sheetName,
@@ -58,10 +70,14 @@ test.describe.serial("Nursery Transfer Requisition Tests", () => {
     await loginPage.login(module, submodule, formName);
     sideMenu = new SideMenuPage(page);
     await sideMenu.sideMenuBar.waitFor();
+
+    // Update Phase
+    phaseCount++;
+    await setCurrPhase(allPhases[phaseCount]);
   });
 
   // ---------------- Create Test ----------------
-  test("Create Nursery Transfer Requisition", async ({ page, db }) => {
+  test(`Create ${formName}`, async ({ page, db }) => {
     await db.deleteData(deleteSQL, { DocNo: docNo, OU: ou[0] });
 
     const { uiVals } = await NurseryTransferRequisitionCreate(
@@ -82,14 +98,17 @@ test.describe.serial("Nursery Transfer Requisition Tests", () => {
     const dbValues = await db.retrieveData(nurserySQLCommand(formName), {
       DocNo: docNo,
     });
+    if (!dbValues) {
+      throwTestFailMsg("C-DB-NF", formName);
+    }
 
     await validateFormValues(createValues, columns, uiVals);
     await validateDBValues([...uiVals, ou], [...columns, "OU"], dbValues[0]);
   });
 
-  // ---------------- Edit Test ----------------
-  test("Edit Nursery Transfer Requisition", async ({ page, db }) => {
-    const { uiVals } = await NurseryTransferRequisitionEdit(
+  // ---------------- Edit Test (Without Saving) ----------------
+  test(`Edit ${formName} Without Saving`, async ({ page, db }) => {
+    const { uiVals } = await NurseryTransferRequisitionEdit1(
       page,
       sideMenu,
       paths,
@@ -103,13 +122,40 @@ test.describe.serial("Nursery Transfer Requisition Tests", () => {
     const dbValues = await db.retrieveData(nurserySQLCommand(formName), {
       DocNo: docNo,
     });
+    if (!dbValues) {
+      throwTestFailMsg("E1-DB-NF", formName);
+    }
+
+    await validateFormValues(createValues, columns, uiVals);
+    await validateDBValues([...uiVals, ou], [...columns, "OU"], dbValues[0]);
+  });
+
+  // ---------------- Edit Test (With Saving) ----------------
+  test(`Edit ${formName} With Saving`, async ({ page, db }) => {
+    const { uiVals } = await NurseryTransferRequisitionEdit2(
+      page,
+      sideMenu,
+      paths,
+      columns,
+      createValues,
+      editValues,
+      ou,
+      docNo,
+    );
+
+    const dbValues = await db.retrieveData(nurserySQLCommand(formName), {
+      DocNo: docNo,
+    });
+    if (!dbValues) {
+      throwTestFailMsg("E2-DB-NF", formName);
+    }
 
     await validateFormValues(editValues, columns, uiVals);
     await validateDBValues([...uiVals, ou], [...columns, "OU"], dbValues[0]);
   });
 
   // ---------------- Delete Test ----------------
-  test("Delete Nursery Transfer Requisition", async ({ page, db }) => {
+  test(`Delete ${formName}`, async ({ page, db }) => {
     await NurseryTransferRequisitionDelete(
       page,
       sideMenu,
@@ -129,7 +175,9 @@ test.describe.serial("Nursery Transfer Requisition Tests", () => {
 
   //---------------- After All ----------------
   test.afterAll(async ({ db }) => {
-    await db.deleteData(deleteSQL, { DocNo: docNo, OU: ou[0] });
+    if (docNo) await db.deleteData(deleteSQL, { DocNo: docNo, OU: ou[0] });
+
+    await editJson(JsonPath, formName, "");
 
     console.log(`End Running: ${formName}`);
   });
